Apr 21 updates: 1. Added configs (in progress) 2. Ready to reconstruct model (in progress)

+++ PredNet/dataset.py
@@ -7,7 +7,7 @@ class PredictorDataset(Dataset):
         self.data = torch.Tensor(data)
     
     def __getitem__(self, index):
-        return self.data[index, :67], self.data[index, 67:71], self.data[index, 71:72], self.data[index, 72:]
+        return self.data[index, :67], self.data[index, 67:71], self.data[index, 71:72], self.data[index, 72:].long()
 
     def __len__(self):
         return self.data.shape[0]
